refactor: paso el codigo de navegacion a una estructura de clases mas limpia
