+++ app/src/main/java/com/project/agilugr/MainActivity.kt
@@ -5,19 +5,18 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.annotation.RequiresApi
-import com.project.agilugr.ui.navigation.navigation
+import com.project.agilugr.ui.navigation.NavigationDirector
 import com.project.agilugr.ui.theme.AgilUGRTheme
 import kotlin.time.ExperimentalTime
 
 @ExperimentalTime
 class MainActivity : ComponentActivity() {
 
-    // API que vamos a usar para interactuar con la sesion del modo focus
-    // val focus_api_interface: FocusAPI = MockFocusAPI.getMockFocusAPI()
+    // APIs que vamos a consumir para tomar los datos del backend
+    val focus_api = MockFocusAPI.getMockFocusAPI()
 
-    // Distintas views de nuestra aplicacion
-    // val focus_mode_session_view = FocusModeSessionView(this.focus_api_interface)
-    // val focus_mode_selector_view = FocusModeSelector(this.focus_api_interface)
+    // Tomamos el director de navegacion para que lance la interfaz grafica
+    val navigation_director = NavigationDirector(focus_api = focus_api)
 
     // Funcion principal
     @RequiresApi(Build.VERSION_CODES.O)
@@ -29,8 +28,8 @@ class MainActivity : ComponentActivity() {
         setContent {
             AgilUGRTheme {
 
-                // Llamamos a la funcion que nos permite manejar la navegacion por las distintas vistas
-                navigation()
+                // Usamos el director de navegacion para lanzar la interfaz grafica
+                navigation_director.buildNavigationAndStartUI()
             }
         }
     }
